refactor: review-replies 폴더 구조 생성
