+++ src/app.module.ts
@@ -7,6 +7,9 @@ import { ReviewsModule } from './reviews/reviews.module'
 import { typeOrmConfig } from './config/typeorm.config'
 import { MapsModule } from './maps/maps.module'
 import { OpenaiModule } from './openai/openai.module'
+import { ReviewRepliesController } from './review-replies/review-replies.controller';
+import { ReviewRepliesService } from './review-replies/review-replies.service';
+import { ReviewRepliesModule } from './review-replies/review-replies.module';
 
 @Module({
   imports: [
@@ -16,8 +19,11 @@ import { OpenaiModule } from './openai/openai.module'
     EventsModule,
     MapsModule,
     ReviewsModule,
-    OpenaiModule
+    OpenaiModule,
+    ReviewRepliesModule
   ],
+  controllers: [ReviewRepliesController],
+  providers: [ReviewRepliesService],
   // controllers: [AppController, UsersController, StoresController, EventsController, MapsController],
   // providers: [AppService, UsersService, StoresService, EventsService, MapsService],
