+++ src/reviews/reviews.module.ts
@@ -4,14 +4,16 @@ import { ReviewsService } from './reviews.service'
 import { ReviewsController } from './reviews.controller'
 import { Review } from './entites/review.entity'
 import { StoresModule } from 'src/stores/stores.module'
-import { ReviewReply } from './entites/review-reply.entity'
+import { ReviewRepliesModule } from 'src/review-replies/review-replies.module'
 
 @Module({
     imports: [
-        TypeOrmModule.forFeature([Review, ReviewReply]),
+        TypeOrmModule.forFeature([Review]),
         StoresModule,
+        ReviewRepliesModule,
     ],
     providers: [ReviewsService],
     controllers: [ReviewsController],
+    exports: [ReviewsModule]
 })
 export class ReviewsModule { }
